# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Restaurant OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tables:
    get:
      operationId: listTables
      tags: [tables]
      summary: List all tables
      responses:
        "200":
          description: Tables
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Table" }
    post:
      operationId: createTable
      tags: [tables]
      summary: Create a table
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TableInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Table" }
  /tables/{id}:
    patch:
      operationId: updateTable
      tags: [tables]
      summary: Update table status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TableUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Table" }

  # ─── RESERVATIONS ─────────────────────────────────────────────
  /reservations:
    get:
      operationId: listReservations
      tags: [reservations]
      summary: List reservations
      parameters:
        - name: date
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Reservations
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Reservation" }
    post:
      operationId: createReservation
      tags: [reservations]
      summary: Create a reservation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReservationInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Reservation" }
  /reservations/{id}:
    get:
      operationId: getReservation
      tags: [reservations]
      summary: Get a reservation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Reservation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Reservation" }
    patch:
      operationId: updateReservation
      tags: [reservations]
      summary: Update a reservation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReservationUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Reservation" }
    delete:
      operationId: deleteReservation
      tags: [reservations]
      summary: Cancel a reservation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── PRODUCTS ─────────────────────────────────────────────────
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List staff members
      parameters:
        - name: role
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Staff list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/StaffMember" }
    post:
      operationId: createStaff
      tags: [staff]
      summary: Add a staff member
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StaffInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StaffMember" }
  /staff/{id}:
    patch:
      operationId: updateStaff
      tags: [staff]
      summary: Update a staff member
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StaffUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StaffMember" }
    delete:
      operationId: deleteStaff
      tags: [staff]
      summary: Remove a staff member
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── SHIFTS ───────────────────────────────────────────────────
  /shifts:
    get:
      operationId: listShifts
      tags: [shifts]
      summary: List shifts
      parameters:
        - name: date
          in: query
          schema: { type: string }
        - name: staffId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Shifts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Shift" }
    post:
      operationId: createShift
      tags: [shifts]
      summary: Create a shift
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ShiftInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Shift" }
  /shifts/{id}:
    patch:
      operationId: updateShift
      tags: [shifts]
      summary: Update a shift
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ShiftUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Shift" }
    delete:
      operationId: deleteShift
      tags: [shifts]
      summary: Delete a shift
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── TASKS ────────────────────────────────────────────────────
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: staffId
          in: query
          schema: { type: integer }
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Task" }
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TaskInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TaskUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── DASHBOARD ────────────────────────────────────────────────
...
components:
  schemas:
...
    Table:
      type: object
      required: [id, number, capacity, status, section]
      properties:
        id: { type: integer }
        number: { type: integer }
        capacity: { type: integer }
        status: { type: string }
        section: { type: string }
        notes: { type: ["string", "null"] }
    TableInput:
      type: object
      required: [number, capacity, section]
      properties:
        number: { type: integer }
        capacity: { type: integer }
        section: { type: string }
        notes: { type: string }
    TableUpdate:
      type: object
      properties:
        status: { type: string }
        capacity: { type: integer }
        section: { type: string }
        notes: { type: string }

    # ── Reservation ──
    Reservation:
      type: object
      required: [id, customerName, customerPhone, partySize, reservedAt, status, createdAt]
      properties:
        id: { type: integer }
        customerId: { type: ["integer", "null"] }
        customerName: { type: string }
        customerPhone: { type: string }
        tableId: { type: ["integer", "null"] }
        partySize: { type: integer }
        reservedAt: { type: string }
        status: { type: string }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
    ReservationInput:
      type: object
      required: [customerName, customerPhone, partySize, reservedAt]
      properties:
        customerId: { type: integer }
        customerName: { type: string, minLength: 1 }
        customerPhone: { type: string, minLength: 1 }
        tableId: { type: integer }
        partySize: { type: integer }
        reservedAt: { type: string }
        notes: { type: string }
    ReservationUpdate:
      type: object
      properties:
        tableId: { type: integer }
        partySize: { type: integer }
        reservedAt: { type: string }
        status: { type: string }
        notes: { type: string }

    # ── Product ──
...
    StaffMember:
      type: object
      required: [id, name, role, phone, status, hireDate, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        role: { type: string }
        phone: { type: string }
        email: { type: ["string", "null"] }
        status: { type: string }
        hireDate: { type: string }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
    StaffInput:
      type: object
      required: [name, role, phone, hireDate]
      properties:
        name: { type: string, minLength: 1 }
        role: { type: string }
        phone: { type: string }
        email: { type: string }
        hireDate: { type: string }
        notes: { type: string }
    StaffUpdate:
      type: object
      properties:
        name: { type: string }
        role: { type: string }
        phone: { type: string }
        email: { type: string }
        status: { type: string }
        notes: { type: string }

    # ── Shift ──
    Shift:
      type: object
      required: [id, staffId, staffName, date, startTime, endTime, role, createdAt]
      properties:
        id: { type: integer }
        staffId: { type: integer }
        staffName: { type: string }
        date: { type: string }
        startTime: { type: string }
        endTime: { type: string }
        role: { type: string }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
    ShiftInput:
      type: object
      required: [staffId, date, startTime, endTime, role]
      properties:
        staffId: { type: integer }
        date: { type: string }
        startTime: { type: string }
        endTime: { type: string }
        role: { type: string }
        notes: { type: string }
    ShiftUpdate:
      type: object
      properties:
        date: { type: string }
        startTime: { type: string }
        endTime: { type: string }
        role: { type: string }
        notes: { type: string }

    # ── Task ──
    Task:
      type: object
      required: [id, title, status, priority, createdAt]
      properties:
        id: { type: integer }
        staffId: { type: ["integer", "null"] }
        staffName: { type: ["string", "null"] }
        title: { type: string }
        description: { type: ["string", "null"] }
        status: { type: string }
        priority: { type: string }
        dueDate: { type: ["string", "null"] }
        createdAt: { type: string }
    TaskInput:
      type: object
      required: [title, priority]
      properties:
        staffId: { type: integer }
        title: { type: string, minLength: 1 }
        description: { type: string }
        priority: { type: string }
        dueDate: { type: string }
    TaskUpdate:
      type: object
      properties:
        staffId: { type: integer }
        title: { type: string }
        description: { type: string }
        status: { type: string }
        priority: { type: string }
        dueDate: { type: string }

    # ── Dashboard ──
...