# lib/api-zod/src/generated/types/staffInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Restaurant OS API
 * OpenAPI spec version: 0.1.0
 */

export interface StaffInput {
  /** @minLength 1 */
  name: string;
  role: string;
  phone: string;
  email?: string;
  hireDate: string;
  notes?: string;
}